backend/requirements.txt
# FastAPI e dependências web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Cache e Queue
redis==5.0.1
celery==5.3.4

# Security
cryptography==41.0.8
bcrypt==4.0.1

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.14.0
transformers==4.35.2
spacy==3.7.2

# NLP específico para português
https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0-py3-none-any.whl

# Monitoring e Observabilidade
prometheus-client==0.19.0
structlog==23.2.0

# File Processing
reportlab==4.0.7
matplotlib==3.8.2
plotly==5.17.0
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
bandit==1.7.5
safety==2.3.5

# Development
pre-commit==3.5.0

# Vulnerability Scanning
python-nmap==0.7.1

# Email
fastapi-mail==1.4.1

# File Storage
minio==7.2.0

# MLflow for ML model management
mlflow==2.8.1

# Hyperparameter optimization
optuna==3.4.0

# Additional utilities
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
typer==0.9.0